Extract todo markdown renderer
